// front/services/models/loop-step-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ConditionJsonld } from './condition-jsonld';

/**
 * 
 * @export
 * @interface LoopStepJsonld
 */
export interface LoopStepJsonld {
    /**
     * 
     * @type {string}
     * @memberof LoopStepJsonld
     */
    'scenario': string;
    /**
     * 
     * @type {string}
     * @memberof LoopStepJsonld
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof LoopStepJsonld
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof LoopStepJsonld
     */
    'stepNumber': number;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof LoopStepJsonld
     */
    'variables'?: { [key: string]: string; };
    /**
     * 
     * @type {boolean}
     * @memberof LoopStepJsonld
     */
    'runAfterFailure'?: boolean;
    /**
     * 
     * @type {Array<ConditionJsonld>}
     * @memberof LoopStepJsonld
     */
    'conditions'?: Array<ConditionJsonld>;
    /**
     * 
     * @type {number}
     * @memberof LoopStepJsonld
     */
    'stepToReturn'?: number;
}