// front/services/models/sleep-step-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SleepStepJsonld
 */
export interface SleepStepJsonld {
    /**
     * 
     * @type {string}
     * @memberof SleepStepJsonld
     */
    'scenario': string;
    /**
     * 
     * @type {string}
     * @memberof SleepStepJsonld
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SleepStepJsonld
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof SleepStepJsonld
     */
    'stepNumber': number;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof SleepStepJsonld
     */
    'variables'?: { [key: string]: string; };
    /**
     * 
     * @type {boolean}
     * @memberof SleepStepJsonld
     */
    'runAfterFailure'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof SleepStepJsonld
     */
    'duration'?: number;
}